Rework code distribution of test management
- add a way to scan test infrastructure for commandline options and
  their defaults
- move most options out of `testing.js` into the modules that work with
  these options
- options used by instance.js and instance-manager.js will end in
  instance-manager.js
- introduce API function to add documentation including duplicate
  checking
- remove redundand options from testsuites
- locate and fix duplicates
- move portManager into its own file
