+++ js/client/modules/@arangodb/testsuites/audit.js
@@ -108,6 +108,6 @@ exports.setup = function (testFns, opts, fnDocs, optionsDoc, allTestPaths) {
   // only enable them in Enterprise Edition
   opts['skipAudit'] = !isEnterprise();
 
-  for (var attrname in functionsDocumentation) { fnDocs[attrname] = functionsDocumentation[attrname]; }
-  for (var i = 0; i < optionsDocumentation.length; i++) { optionsDoc.push(optionsDocumentation[i]); }
+  tu.CopyIntoObject(fnDocs, functionsDocumentation);
+  tu.CopyIntoList(optionsDoc, optionsDocumentation);
 };
