+++ js/client/modules/@arangodb/testsuites/chaos.js
@@ -29,7 +29,6 @@ const functionsDocumentation = {
   'chaos': 'chaos tests',
   'deadlock': 'deadlock tests'
 };
-const optionsDocumentation = [];
 
 const _ = require('lodash');
 const tu = require('@arangodb/testutils/test-utils');
@@ -111,6 +110,5 @@ exports.setup = function (testFns, opts, fnDocs, optionsDoc, allTestPaths) {
   Object.assign(allTestPaths, testPaths);
   testFns['chaos'] = chaos;
   testFns['deadlock'] = deadlock;
-  for (var attrname in functionsDocumentation) { fnDocs[attrname] = functionsDocumentation[attrname]; }
-  for (var i = 0; i < optionsDocumentation.length; i++) { optionsDoc.push(optionsDocumentation[i]); }
+  tu.CopyIntoObject(fnDocs, functionsDocumentation);
 };
