+++ js/client/modules/@arangodb/testsuites/communication.js
@@ -29,7 +29,6 @@ const functionsDocumentation = {
   'communication': 'communication tests',
   'communication_ssl': 'communication tests with SSL'
 };
-const optionsDocumentation = [];
 
 const _ = require('lodash');
 const tu = require('@arangodb/testutils/test-utils');
@@ -63,6 +62,5 @@ exports.setup = function (testFns, opts, fnDocs, optionsDoc, allTestPaths) {
   testFns['communication'] = communication;
   testFns['communication_ssl'] = communicationSsl;
 
-  for (var attrname in functionsDocumentation) { fnDocs[attrname] = functionsDocumentation[attrname]; }
-  for (var i = 0; i < optionsDocumentation.length; i++) { optionsDoc.push(optionsDocumentation[i]); }
+  tu.CopyIntoObject(fnDocs, functionsDocumentation);
 };
