+++ js/client/modules/@arangodb/testsuites/drivers.js
@@ -166,6 +166,6 @@ exports.setup = function (testFns, opts, fnDocs, optionsDoc, allTestPaths) {
   opts.driverScript = 'run_tests.sh';
   opts.driverScriptInterpreter = '/bin/bash';
   testFns['driver'] = driver;
-  for (var attrname in functionsDocumentation) { fnDocs[attrname] = functionsDocumentation[attrname]; }
-  for (var i = 0; i < optionsDocumentation.length; i++) { optionsDoc.push(optionsDocumentation[i]); }
+  tu.CopyIntoObject(fnDocs, functionsDocumentation);
+  tu.CopyIntoList(optionsDoc, optionsDocumentation);
 };
