+++ js/client/modules/@arangodb/testsuites/fuerte.js
@@ -28,12 +28,10 @@
 const functionsDocumentation = {
   'fuerte': 'fuerte gtest test suites'
 };
-const optionsDocumentation = [
-  '   - `skipFuerte`: if set to true the fuerte tests are skipped'
-];
 
 const fs = require('fs');
 const pu = require('@arangodb/testutils/process-utils');
+const tu = require('@arangodb/testutils/test-utils');
 const im = require('@arangodb/testutils/instance-manager');
 const {getGTestResults} = require('@arangodb/testutils/result-processing');
 
@@ -66,9 +64,6 @@ function gtestRunner(options) {
   let testResultJsonFile = fs.join(rootDir, 'testResults.json');
 
   const run = locateGTest('fuertetest');
-  if (options.skipFuerte) {
-    return results;
-  }
 
   if (run === '') {
     results.failed += 1;
@@ -134,8 +129,5 @@ exports.setup = function (testFns, opts, fnDocs, optionsDoc, allTestPaths) {
   Object.assign(allTestPaths, testPaths);
   testFns['fuerte'] = gtestRunner;
 
-  opts['skipFuerte'] = false;
-
-  for (var attrname in functionsDocumentation) { fnDocs[attrname] = functionsDocumentation[attrname]; }
-  for (var i = 0; i < optionsDocumentation.length; i++) { optionsDoc.push(optionsDocumentation[i]); }
+  tu.CopyIntoObject(fnDocs, functionsDocumentation);
 };
