+++ js/client/modules/@arangodb/testsuites/loadBalancing.js
@@ -29,9 +29,6 @@
 const functionsDocumentation = {
   'load_balancing': 'load balancing tests'
 };
-const optionsDocumentation = [
-  '   - `skipLoadBalancing : testing load_balancing will be skipped.'
-];
 
 const _ = require('lodash');
 const tu = require('@arangodb/testutils/test-utils');
@@ -55,16 +52,6 @@ const testPaths = {
 ////////////////////////////////////////////////////////////////////////////////
 
 function loadBalancingClient (options) {
-  if (options.skipLoadBalancing === true) {
-    print('skipping Load Balancing tests!');
-    return {
-      load_balancing: {
-        status: true,
-        skipped: true
-      }
-    };
-  }
-
   print(CYAN + 'Load Balancing tests...' + RESET);
   const excludeAuth = (fn) => { return (fn.indexOf('-auth') === -1); };
   let testCases = tu.scanTestPaths(testPaths.load_balancing, options)
@@ -87,16 +74,6 @@ function loadBalancingClient (options) {
 ////////////////////////////////////////////////////////////////////////////////
 
 function loadBalancingAuthClient (options) {
-  if (options.skipLoadBalancing === true) {
-    print('skipping Load Balancing tests!');
-    return {
-      load_balancing_auth: {
-        status: true,
-        skipped: true
-      }
-    };
-  }
-
   print(CYAN + 'Load Balancing with Authentication tests...' + RESET);
   const excludeNoAuth = (fn) => { return (fn.indexOf('-noauth') === -1); };
   let testCases = tu.scanTestPaths(testPaths.load_balancing, options)
@@ -119,8 +96,5 @@ exports.setup = function (testFns, opts, fnDocs, optionsDoc, allTestPaths) {
   testFns['load_balancing'] = loadBalancingClient;
   testFns['load_balancing_auth'] = loadBalancingAuthClient;
 
-  opts['skipLoadBalancing'] = false;
-
-  for (var attrname in functionsDocumentation) { fnDocs[attrname] = functionsDocumentation[attrname]; }
-  for (var i = 0; i < optionsDocumentation.length; i++) { optionsDoc.push(optionsDocumentation[i]); }
+  tu.CopyIntoObject(fnDocs, functionsDocumentation);
 };
