+++ js/client/modules/@arangodb/testsuites/replication.js
@@ -226,6 +226,6 @@ exports.setup = function (testFns, opts, fnDocs, optionsDoc, allTestPaths) {
   testFns['replication_static'] = replicationStatic;
   testFns['replication_sync'] = replicationSync;
   testFns['http_replication'] = shellClientReplicationApi;
-  for (var attrname in functionsDocumentation) { fnDocs[attrname] = functionsDocumentation[attrname]; }
-  for (var i = 0; i < optionsDocumentation.length; i++) { optionsDoc.push(optionsDocumentation[i]); }
+  tu.CopyIntoObject(fnDocs, functionsDocumentation);
+  tu.CopyIntoList(optionsDoc, optionsDocumentation);
 };
