+++ js/client/modules/@arangodb/testutils/crash-utils.js
@@ -28,6 +28,8 @@
 const fs = require('fs');
 const yaml = require('js-yaml');
 const internal = require('internal');
+const tu = require('@arangodb/testutils/test-utils');
+const versionHas = require("@arangodb/test-helper").versionHas;
 const {
   executeExternal,
   executeExternalAndWait,
@@ -257,350 +259,6 @@ function generateCoreDumpGDB (instanceInfo, options, storeArangodPath, pid, gene
 }
 
 // //////////////////////////////////////////////////////////////////////////////
-// / @brief analyzes a core dump using lldb (macos)
-// /
-// / We assume the system has core files in /cores/, and we have a lldb.
-// //////////////////////////////////////////////////////////////////////////////
-
-function analyzeCoreDumpMac (instanceInfo, options, storeArangodPath, pid) {
-  let lldbOutputFile = fs.getTempFile();
-
-  let command;
-  command = '(';
-  command += 'printf \'bt \n\n';
-  // LLDB doesn't have an equivilant of `bt full` so we try to show the upper
-  // most 5 frames with all variables
-  for (var i = 0; i < 5; i++) {
-    command += 'frame variable\\n up \\n';
-  }
-  command += ' thread backtrace all\\n\';';
-  command += 'sleep 10;';
-  command += 'echo quit;';
-  command += 'sleep 2';
-  command += ') | lldb ' + storeArangodPath;
-  command += ' -c /cores/core.' + pid;
-  command += ' > ' + lldbOutputFile + ' 2>&1';
-  const args = ['-c', command];
-  print("launching LLDB in foreground: " + JSON.stringify(args));
-
-  sleep(5);
-  executeExternalAndWait('/bin/bash', args);
-  GDB_OUTPUT += `--------------------------------------------------------------------------------
-Crash analysis of: ` + JSON.stringify(instanceInfo.getStructure()) + '\n\n';
-  let thisDump = fs.read(lldbOutputFile);
-  GDB_OUTPUT += thisDump;
-  if (options.extremeVerbosity === true) {
-    print(thisDump);
-  }
-  return 'lldb ' + storeArangodPath + ' -c /cores/core.' + pid;
-}
-
-// //////////////////////////////////////////////////////////////////////////////
-// / @brief analyzes a core dump using lldb (macos)
-// /
-// / We assume the system has core files in /cores/, and we have a lldb.
-// //////////////////////////////////////////////////////////////////////////////
-
-function generateCoreDumpMac (instanceInfo, options, storeArangodPath, pid, generateCoreDump) {
-  let lldbOutputFile = fs.getTempFile();
-  let gcore = '';
-  if (generateCoreDump) {
-    if (options.coreDirectory === '') {
-      gcore = ` process save-core core.${instanceInfo.pid}\\n`;
-    } else {
-      gcore = ` process save-core ${options.coreDirectory}/core.${instanceInfo.pid}\\n`;
-    }
-  }
-
-  let command;
-  command = '(';
-  command += 'printf \'bt \n\n';
-  // LLDB doesn't have an equivilant of `bt full` so we try to show the upper
-  // most 5 frames with all variables
-  for (var i = 0; i < 5; i++) {
-    command += 'frame variable\\n up \\n';
-  }
-  command += ` thread backtrace all\\n`;
-  command += gcore;
-  command += ` kill\\n';`;
-  command += 'sleep 10;';
-  command += 'echo quit;';
-  command += 'sleep 2';
-  command += ') | lldb ';
-  command += ` --attach-pid ${pid} `;
-  command += storeArangodPath;
-  command += ' > ' + lldbOutputFile + ' 2>&1';
-  const args = ['-c', command];
-  print("launching LLDB in background: " + JSON.stringify(command));
-  return {
-    pid: executeExternal('/bin/bash', args),
-    file: lldbOutputFile,
-    hint:  'lldb ' + storeArangodPath + ' -c /cores/core.' + pid,
-    verbosePrint: true
-  };
-}
-
-// //////////////////////////////////////////////////////////////////////////////
-// / @brief analyzes a core dump using cdb (Windows)
-// /  cdb is part of the WinDBG package.
-// //////////////////////////////////////////////////////////////////////////////
-
-
-// //////////////////////////////////////////////////////////////////////////////
-// / @brief check whether process does bad on the wintendo
-// //////////////////////////////////////////////////////////////////////////////
-
-function runProcdump (options, instanceInfo, rootDir, pid, instantDump = false) {
-  let procdumpArgs = [ ];
-  let dumpFile = fs.join(rootDir, 'core_' + pid + '.dmp');
-  if (options.exceptionFilter != null) {
-    procdumpArgs = [
-      '-accepteula',
-      '-64',
-    ];
-    if (!instantDump) {
-      procdumpArgs.push('-e');
-      procdumpArgs.push(options.exceptionCount);
-    }
-    let filters = options.exceptionFilter.split(',');
-    for (let which in filters) {
-      procdumpArgs.push('-f');
-      procdumpArgs.push(filters[which]);
-    }
-    procdumpArgs.push('-ma');
-    procdumpArgs.push(pid);
-    procdumpArgs.push(dumpFile);
-  } else {
-    procdumpArgs = [
-      '-accepteula',
-    ];
-    if (!instantDump) {
-      procdumpArgs.push('-e');
-    }
-    procdumpArgs.push('-ma');
-    procdumpArgs.push(pid);
-    procdumpArgs.push(dumpFile);
-  }
-  try {
-    if (options.extremeVerbosity) {
-      print(Date() + " Starting procdump: " + JSON.stringify(procdumpArgs));
-    }
-    instanceInfo.coreFilePattern = dumpFile;
-    if (instantDump) {
-      // Wait for procdump to have written the dump before we attempt to kill the process:
-      instanceInfo.monitor = executeExternalAndWait('procdump', procdumpArgs);
-    } else {
-      instanceInfo.monitor = executeExternal('procdump', procdumpArgs);
-      // try to give procdump a little time to catch up with the process
-      sleep(0.25);
-      let status = statusExternal(instanceInfo.monitor.pid, false);
-      if (status.hasOwnProperty('signal')) {
-        print(RED + 'procdump didn\'t come up: ' + JSON.stringify(status));
-        instanceInfo.monitor.status = status;
-        return false;
-      }
-    }
-  } catch (x) {
-    print(Date() + ' failed to start procdump - is it installed?');
-    // throw x;
-    return false;
-  }
-  return true;
-}
-
-function stopProcdump (options, instanceInfo, force = false) {
-  if (instanceInfo.hasOwnProperty('monitor') &&
-      instanceInfo.monitor.pid !== null) {
-    if (force) {
-      print(Date() + " sending TERM to procdump to make it exit");
-      instanceInfo.monitor.status = killExternal(instanceInfo.monitor.pid, termSignal);
-    } else {
-      print(Date() + " waiting for procdump to exit");
-      statusExternal(instanceInfo.monitor.pid, true);
-    }
-    instanceInfo.monitor.pid = null;
-  }
-}
-
-function calculateMonitorValues(options, instanceInfo, pid, cmd) {}
-
-function isEnabledWindowsMonitor(options, instanceInfo, pid, cmd) {
-  return false;
-}
-
-function readCdbFileFiltered(cdbOutputFile) {
-  try {
-    const filters = JSON.parse(fs.read(
-      fs.join(pu.JS_DIR,
-              'client/modules/@arangodb/testutils',
-              'filter_cdb_stacks.json')));
-    const buf = fs.readBuffer(cdbOutputFile);
-    let lineStart = 0;
-    let maxBuffer = buf.length;
-    let inStack = false;
-    let stack = [];
-    let longStack = [];
-    for (let j = 0; j < maxBuffer; j++) {
-      if (buf[j] === 10) { // \n
-        var line = buf.asciiSlice(lineStart, j);
-        lineStart = j + 1;
-
-        if (line.search('RetAddr') === 0) {
-          inStack = true;
-        }
-        if (inStack) {
-          if ((line.search(' 000') !== -1) ||
-              (line.search('--------') !== -1)) {
-            // cut off left part with addresses:
-            line = line.substring(98);
-            longStack.push(line);
-            // cut off arguments only for filter string::
-            let plusPos = line.search('\\+');
-            if (plusPos > 0) {
-              line = line.substring(0, plusPos);
-            }
-            if (line.length > 1) {
-              stack.push(line);
-            }
-          }
-          if (line.length === 0) {
-            if (!filterStack(stack, filters)) {
-              print("did not filter this stack: ");
-              print(stack);
-              longStack.forEach(line => {
-                GDB_OUTPUT += line.trim() + '\n';
-              });
-              GDB_OUTPUT += '\n';
-            }
-            
-            stack = [];
-            longStack = [];
-            inStack = false;
-          }
-        } else {
-          if (line.search('NatVis') === -1) {
-            GDB_OUTPUT += line.trim() + '\n';
-          }
-        }
-      }
-    }
-  } catch (ex) {
-    let err="failed to read " + cdbOutputFile + " -> " + ex + '\n' + ex.stack;
-    print(err);
-  }
-}
-
-const core_rx = new RegExp('core_.*_.*.dmp');
-function analyzeCoreDumpWindows (instanceInfo) {
-  let cdbOutputFile = fs.getTempFile();
-
-  if (!fs.exists(instanceInfo.coreFilePattern)) {
-    const m = 'core file ' + instanceInfo.coreFilePattern + ' not found?';
-    instanceInfo.coreFilePattern = null;
-    // locate arangods crash handler self generated minidump files:
-    fs.list(instanceInfo.coreDirectory).forEach(file => {
-      if (file.match(core_rx) !== null) {
-        instanceInfo.coreFilePattern = fs.join(instanceInfo.coreDirectory, file);
-      }
-    });
-    if (instanceInfo.coreDirectory === null) {
-      print(m);
-      return;
-    }
-  }
-  print('found minidump file: ' + instanceInfo.coreFilePattern);
-
-  const dbgCmds = [
-    '.logopen ' + cdbOutputFile,
-    '!analyze -v', // print verbose analysis
-    'dv', // analyze local variables (if)
-    '~*kb', // print all threads stack traces
-    'q' // quit the debugger
-  ];
-
-  const args = [
-    '-z',
-    instanceInfo.coreFilePattern,
-    '-lines',
-    '-logo',
-    cdbOutputFile,
-    '-c',
-    dbgCmds.join('; ')
-  ];
-
-  sleep(5);
-  print('running cdb in foreground: ' + JSON.stringify(args));
-  process.env['_NT_DEBUG_LOG_FILE_OPEN'] = cdbOutputFile;
-  executeExternalAndWait('cdb', args);
-  GDB_OUTPUT += `--------------------------------------------------------------------------------
-Crash analysis of: ` + JSON.stringify(instanceInfo.getStructure()) + '\n\n';
-  // cdb will output to stdout anyways, so we can't turn this off here.
-  readCdbFileFiltered(cdbOutputFile);
-  return 'cdb ' + args.join(' ');
-}
-
-function generateCoreDumpWindows (instanceInfo) {
-  let cdbOutputFile = fs.getTempFile();
-  let dbgCmds = [
-    '.logopen ' + cdbOutputFile,
-    '!analyze -v', // print verbose analysis
-    'dv', // analyze local variables (if)
-    '~*kb', // print all threads stack traces
-    `.dump /mFhtipcy ${instanceInfo.coreFilePattern}`, // write a compact dump file
-    '.kill', // terminate the debugged process
-    'q' // quit the debugger
-  ];
-
-  const args = [
-    '-p',
-    instanceInfo.pid,
-    '-lines',
-    '-logo',
-    cdbOutputFile,
-    '-c',
-    dbgCmds.join('; ')
-  ];
-
-  print('running cdb in background: ' + JSON.stringify(args));
-  process.env['_NT_DEBUG_LOG_FILE_OPEN'] = cdbOutputFile;
-  return {
-    pid: executeExternal('cdb', args),
-    file: cdbOutputFile,
-    hint: 'cdb ' + args.join(' '),
-    verbosePrint: false
-  };
-}
-
-function checkMonitorAlive (binary, instanceInfo, options, res) {
-  if (instanceInfo.hasOwnProperty('monitor') ) {
-    // Windows: wait for procdump to do its job...
-    if (!instanceInfo.monitor.hasOwnProperty('status')) {
-      let rc = statusExternal(instanceInfo.monitor.pid, false);
-      if (rc.status !== 'RUNNING') {
-        print(RED + "Procdump of " + instanceInfo.pid + " is gone: " + rc + RESET);
-        instanceInfo.monitor = rc;
-        // procdump doesn't set proper exit codes, check for
-        // dumps that may exist:
-        if (fs.exists(instanceInfo.coreFilePattern)) {
-          print("checkMonitorAlive: marking crashy");
-          instanceInfo.monitor.monitorExited = true;
-          instanceInfo.monitor.pid = null;
-          pu.serverCrashed = true;
-          options.cleanup = false;
-          instanceInfo['exitStatus'] = {};
-          analyzeCrash(binary, instanceInfo, options, "the process monitor commanded error");
-          Object.assign(instanceInfo.exitStatus,
-                        killExternal(instanceInfo.pid, abortSignal));
-          return false;
-        }
-      }
-    }
-    else return instanceInfo.monitor.exitStatus;
-  }
-  return true;
-}
-
-// //////////////////////////////////////////////////////////////////////////////
 // / @brief the bad has happened, tell it the user and try to gather more
 // /        information about the incident.
 // //////////////////////////////////////////////////////////////////////////////
@@ -683,11 +341,7 @@ function analyzeCrash (binary, instanceInfo, options, checkStr) {
   sleep(5);
 
   let hint = '';
-  if (platform === 'darwin') {
-    hint = analyzeCoreDumpMac(instanceInfo, options, binary, instanceInfo.pid);
-  } else {
-    hint = analyzeCoreDump(instanceInfo, options, binary, instanceInfo.pid);
-  }
+  hint = analyzeCoreDump(instanceInfo, options, binary, instanceInfo.pid);
   instanceInfo.exitStatus.gdbHint = 'Run debugger with "' + hint + '"';
 
 }
@@ -711,9 +365,6 @@ function generateCrashDump (binary, instanceInfo, options, checkStr) {
   if (options.test !== undefined) {
     print(CYAN + instanceInfo.name + " - in single test mode, hard killing." + RESET);
     instanceInfo.exitStatus = killExternal(instanceInfo.pid, termSignal);
-  } else if (platform === 'darwin') {
-    instanceInfo.debuggerInfo = generateCoreDumpMac(instanceInfo, options, binary, instanceInfo.pid, generateCoreDump);
-    instanceInfo.exitStatus = { status: 'TERMINATED'};
   } else {
     instanceInfo.debuggerInfo = generateCoreDumpGDB(instanceInfo, options, binary, instanceInfo.pid, generateCoreDump);
     instanceInfo.exitStatus = { status: 'TERMINATED'};
@@ -758,36 +409,29 @@ function aggregateDebugger(instanceInfo, options) {
 --------------------------------------------------------------------------------
 Crash analysis of: ` + JSON.stringify(instanceInfo.getStructure()) + '\n\n';
 
-  if (platform === 'darwin') {
-    const buf = fs.readBuffer(instanceInfo.debuggerInfo.file);
-    let lineStart = 0;
-    let maxBuffer = buf.length;
-
-    for (let j = 0; j < maxBuffer; j++) {
-      if (buf[j] === 10) { // \n
-        const line = buf.asciiSlice(lineStart, j);
-        lineStart = j + 1;
-        if (line.search('bytes of data for memory region at') !== -1) {
-          continue;
-        }
-        GDB_OUTPUT += line + '\n';
-        if (options.extremeVerbosity === true && instanceInfo.debuggerInfo.verbosePrint) {
-          print(line);
-        }
-      }
-    }
-  } else {
-    readGdbFileFiltered(instanceInfo.debuggerInfo.file);
-  }
+  readGdbFileFiltered(instanceInfo.debuggerInfo.file);
   return instanceInfo.debuggerInfo.hint;
 }
 
 exports.aggregateDebugger = aggregateDebugger;
 exports.generateCrashDump = generateCrashDump;
-exports.checkMonitorAlive = checkMonitorAlive;
 exports.analyzeCrash = analyzeCrash;
-exports.runProcdump = runProcdump;
-exports.stopProcdump = stopProcdump;
-exports.isEnabledWindowsMonitor = isEnabledWindowsMonitor;
-exports.calculateMonitorValues = calculateMonitorValues;
 Object.defineProperty(exports, 'GDB_OUTPUT', { get: () => GDB_OUTPUT, set: (value) => { GDB_OUTPUT = value; }});
+exports.registerOptions = function(optionsDefaults, optionsDocumentation) {
+  const isSan = versionHas('asan') || versionHas('tsan');
+  tu.CopyIntoObject(optionsDefaults, {
+    'coreCheck': false,
+    'coreAbort': false,
+    'coreDirectory': '/var/tmp',
+    'coreGen': !isSan,
+  });
+
+  tu.CopyIntoList(optionsDocumentation, [
+    ' Crash analysis related:',
+    '   - `coreGen`: whether debuggers should generate a coredump after getting stacktraces',
+    '   - `coreAbort`: if we should use sigAbrt in order to terminate process instead of GDB attaching',
+    '   - `coreCheck`: if set to true, we will attempt to locate a coredump to ',
+    '                  produce a backtrace in the event of a crash',
+    '',
+  ]);
+};
