+++ js/client/modules/@arangodb/testutils/process-utils.js
@@ -30,6 +30,7 @@
 const _ = require('lodash');
 const fs = require('fs');
 const rp = require('@arangodb/testutils/result-processing');
+const tu = require('@arangodb/testutils/test-utils');
 const yaml = require('js-yaml');
 const internal = require('internal');
 const crashUtils = require('@arangodb/testutils/crash-utils');
@@ -126,12 +127,12 @@ function coverageEnvironment () {
 // / required to be called first.
 // //////////////////////////////////////////////////////////////////////////////
 
-function setupBinaries (builddir, buildType, configDir) {
-  if (builddir === '') {
+function setupBinaries (options) {
+  if (options.build === '') {
     if (fs.exists('build') && fs.exists(fs.join('build', 'bin'))) {
-      builddir = 'build';
+      options.build = 'build';
     } else if (fs.exists('bin')) {
-      builddir = '.';
+      options.build = '.';
     } else {
       print('FATAL: cannot find binaries, use "--build"\n');
 
@@ -141,7 +142,7 @@ function setupBinaries (builddir, buildType, configDir) {
     }
   }
 
-  BIN_DIR = fs.join(builddir, 'bin');
+  BIN_DIR = fs.join(options.build, 'bin');
   if (!fs.exists(BIN_DIR)) {
     BIN_DIR = fs.join(TOP_DIR, BIN_DIR);
   }
@@ -150,18 +151,9 @@ function setupBinaries (builddir, buildType, configDir) {
     BIN_DIR = fs.join(TOP_DIR, 'bin');
   }
 
-  UNITTESTS_DIR = fs.join(fs.join(builddir, 'tests'));
+  UNITTESTS_DIR = fs.join(fs.join(options.build, 'tests'));
   if (!fs.exists(UNITTESTS_DIR)) {
     UNITTESTS_DIR = fs.join(TOP_DIR, UNITTESTS_DIR);
-  }
-
-  if (buildType !== '') {
-    if (fs.exists(fs.join(BIN_DIR, buildType))) {
-      BIN_DIR = fs.join(BIN_DIR, buildType);
-    }
-    if (fs.exists(fs.join(UNITTESTS_DIR, buildType))) {
-      UNITTESTS_DIR = fs.join(UNITTESTS_DIR, buildType);
-    }
   }
 
   ARANGOBACKUP_BIN = fs.join(BIN_DIR, 'arangobackup' + executableExt);
@@ -174,13 +166,13 @@ function setupBinaries (builddir, buildType, configDir) {
   ARANGOSH_BIN = fs.join(BIN_DIR, 'arangosh' + executableExt);
   ARANGO_SECURE_INSTALLATION_BIN = fs.join(BIN_DIR, 'arango-secure-installation' + executableExt);
 
-  CONFIG_ARANGODB_DIR = fs.join(builddir, 'etc', 'arangodb3');
+  CONFIG_ARANGODB_DIR = fs.join(options.build, 'etc', 'arangodb3');
   if (!fs.exists(CONFIG_ARANGODB_DIR)) {
     CONFIG_ARANGODB_DIR = fs.join(TOP_DIR, CONFIG_ARANGODB_DIR);
   }
 
   CONFIG_RELATIVE_DIR = fs.join(TOP_DIR, 'etc', 'relative');
-  CONFIG_DIR = fs.join(TOP_DIR, configDir);
+  CONFIG_DIR = fs.join(TOP_DIR, options.configDir);
 
   JS_DIR = fs.join(TOP_DIR, 'js');
   JS_ENTERPRISE_DIR = fs.join(TOP_DIR, 'enterprise/js');
@@ -205,25 +197,6 @@ function setupBinaries (builddir, buildType, configDir) {
   checkFiles.forEach((file) => {
     if (!fs.isFile(file)) {
       throw new Error('unable to locate ' + file);
-    }
-  });
-
-  // fiddle in suppressions for sanitizers if not already set from an
-  // outside script. this can populate the environment variables
-  // - ASAN_OPTIONS
-  // - UBSAN_OPTIONS
-  // - LSAN_OPTIONS
-  // - TSAN_OPTIONS
-  // note: this code repeats existing logic that can be found in
-  // scripts/unittest as well. according to @dothebart it must be
-  // present in both code locations.
-  ["asan", "ubsan", "lsan", "tsan"].forEach((san) => {
-    let envName = san.toUpperCase() + "_OPTIONS";
-    let fileName = san + "_arangodb_suppressions.txt";
-    if (!process.env.hasOwnProperty(envName) &&
-        fs.exists(fileName)) {
-      process.env[envName] = `suppressions=${fs.join(fs.makeAbsolute(''), fileName)}`;
-      print('preparing ' + san + ' environment:', envName + '=' + process.env[envName]);
     }
   });
 
@@ -398,7 +371,6 @@ function executeAndWait (cmd, args, options, valgrindTest, rootDir, coreCheck = 
   let res = executeExternalAndWait(cmd, args, false, timeout * 1000, coverageEnvironment());
   instanceInfo.pid = res.pid;
   instanceInfo.exitStatus = res;
-  crashUtils.calculateMonitorValues(options, instanceInfo, res.pid, cmd);
   const deltaTime = time() - startTime;
 
   let errorMessage = ' - ';
@@ -532,3 +504,17 @@ Object.defineProperty(exports, 'CONFIG_ARANGODB_DIR', {get: () => CONFIG_ARANGOD
 Object.defineProperty(exports, 'CONFIG_RELATIVE_DIR', {get: () => CONFIG_RELATIVE_DIR});
 Object.defineProperty(exports, 'serverCrashed', {get: () => serverCrashedLocal, set: (value) => { serverCrashedLocal = value; } });
 Object.defineProperty(exports, 'serverFailMessages', {get: () => serverFailMessagesLocal, set: (value) => { serverFailMessagesLocal = value; }});
+exports.registerOptions = function(optionsDefaults, optionsDocumentation) {
+  tu.CopyIntoObject(optionsDefaults, {
+    'build': '',
+    'configDir': 'etc/testing',
+  });
+
+  tu.CopyIntoList(optionsDocumentation, [
+    ' Environment options:',
+    '   - `build`: the directory containing the binaries',
+    '   - `configDir`: the directory containing the config files, defaults to',
+    '                  etc/testing',
+    ''
+  ]);
+};
